# Migrate glu from the THC to ATen (CUDA)