--- a/BUILD.bazel
+++ b/BUILD.bazel
@@ -406,7 +406,6 @@ filegroup(
     srcs = [
         "aten/src/THCUNN/BCECriterion.cu.cc",
         "aten/src/THCUNN/ELU.cu.cc",
-        "aten/src/THCUNN/GatedLinearUnit.cu.cc",
         "aten/src/THCUNN/HardTanh.cu.cc",
         "aten/src/THCUNN/LeakyReLU.cu.cc",
         "aten/src/THCUNN/LogSigmoid.cu.cc",