--- a/aten/src/ATen/native/Activation.h
+++ b/aten/src/ATen/native/Activation.h
@@ -15,7 +15,6 @@ using structured_activation_backward_fn = void (*)(TensorIteratorBase&);
 
 using activation_fn = void (*)(TensorIterator&);
 using activation_backward_fn = void (*)(TensorIterator&);
-using glu_backward_cuda_fn = void (*)(TensorIteratorBase& iter, int64_t dim);
 using softplus_fn = void (*)(TensorIteratorBase&, const Scalar&, const Scalar&);
 using softplus_backward_fn = void (*)(TensorIteratorBase&, const Scalar&, const Scalar&);
 using threshold_fn = void (*)(TensorIteratorBase&, const Scalar&, const Scalar&);
@@ -54,7 +53,6 @@ DECLARE_DISPATCH(leaky_relu_fn, leaky_relu_stub);
 DECLARE_DISPATCH(leaky_relu_backward_fn, leaky_relu_backward_stub);
 DECLARE_DISPATCH(activation_fn, glu_stub);
 DECLARE_DISPATCH(activation_backward_fn, glu_backward_stub);
-DECLARE_DISPATCH(glu_backward_cuda_fn, glu_backward_cuda_stub);
 DECLARE_DISPATCH(structured_activation_fn, silu_stub);
 DECLARE_DISPATCH(structured_activation_backward_fn, silu_backward_stub);
 DECLARE_DISPATCH(structured_activation_fn, mish_stub);