… aten/src/ATen/cuda/LegacyTHFunctionsCUDA.cpp
… aten/src/THCUNN/GatedLinearUnit.cu
… aten/src/THCUNN/generic/GatedLinearUnit.cu